useFirestore, createContext, useRoom, useSetup を1本化
