--- src/firestore/useFirestore.1.ts
@@ -1,5 +1,0 @@
-import type { Firestore } from "firebase/firestore"
-import { createContext } from "../createContext.1"
-
-export const [keyWithFirestore, useFirestore] =
-  createContext<Firestore>("useSetup")
