+++ src/firestore/useFirestore.ts
@@ -1,5 +1,5 @@
 import type { Firestore } from "firebase/firestore"
 import { createContext } from "../createContext"
 
-export const [FirestoreProvider, useFirestore] =
-  createContext<Firestore>("FirestoreProvider")
+export const [keyWithFirestore, useFirestore] =
+  createContext<Firestore>("Firestore")
