+++ src/useSetup.ts
@@ -1,13 +1,13 @@
 import { createContext } from "./createContext"
 import type { Room } from "./zod/roomZod"
 
-export function useSetup(roomId: Room["id"]): (() => void) | null {
-  return useImpl()(roomId)
+export function useSetup(
+  roomId: Room["id"]
+): (() => void | PromiseLike<void>) | null {
+  return _useImpl()(roomId)
 }
 
-export { ImplProvider as UseSetupProvider }
-
-const [ImplProvider, useImpl] = createContext<typeof useSetup>(
-  "UseSetupProvider",
+export const [keyWithUseSetup, _useImpl] = createContext<typeof useSetup>(
+  "useSetup",
   () => null
 )
